pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = ""
authors = ["Eagle <ii-sys@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
Flask = "^3.0.3"
gunicorn = "^23.0.0"
psycopg2-binary = "^2.9.10"

[tool.poetry.group.dev.dependencies]
flake8 = "^7.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

